Testing/Performance: * Fix incorrect significant figure handling. * Add support for comments to the performance postprocessing scripts. * Add command line processing utilities to `bench.cu`. * Add --no-header flag and --device flag to `bench.cu`. Bug 200397103 Bug 2011463 git-commit a09a4b9a6e7d93268d4e18517921b394f5d63ffb git-author Bryce Adelstein Lelbach aka wash <brycelelbach@gmail.com> VDVS: http://ausdvs.nvidia.com/Build_Results?virtualId=1000117887&which_page=current_build
[git-p4: depot-paths = "//sw/gpgpu/thrust/": change = 23748376]

+++ internal/benchmark/timer.h
@@ -27,10 +27,10 @@ class timer
     public:
     timer()
     {
-        CUDA_SAFE_CALL(cudaEventCreate(&_start)); 
+        CUDA_SAFE_CALL(cudaEventCreate(&_start));
         CUDA_SAFE_CALL(cudaEventCreate(&_end));
     }
-    
+
     ~timer()
     {
         CUDA_SAFE_CALL(cudaEventDestroy(_start));
@@ -38,26 +38,26 @@ class timer
     }
 
     void start()
-    { 
-        CUDA_SAFE_CALL(cudaEventRecord(_start,0));
+    {
+        CUDA_SAFE_CALL(cudaEventRecord(_start, 0));
     }
-    
+
     void stop()
-    { 
+    {
         CUDA_SAFE_CALL(cudaEventRecord(_end, 0));
         CUDA_SAFE_CALL(cudaEventSynchronize(_end));
     }
 
-    float milliseconds_elapsed()
-    { 
+    double milliseconds_elapsed()
+    {
         float elapsed_time;
         CUDA_SAFE_CALL(cudaEventElapsedTime(&elapsed_time, _start, _end));
         return elapsed_time;
     }
 
-    float seconds_elapsed()
-    { 
-        return milliseconds_elapsed() / 1000.0f;
+    double seconds_elapsed()
+    {
+        return milliseconds_elapsed() / 1000.0;
     }
 };
